Version 3 - Route all power headers on top layer
Previously I had the upper right power rail in the user prototype area
routed on the bottom layer whereas the rest were routed on the bottom.
This commit makes them all the same and makes it easier for the user
to see which pads are connected to what just by looking at the traces
through the top soldermask.