Initial port of MiP-Capi to Arduino
This commit is just a placeholder. It contains a rough port of my
existing C API to the Arduino. All of the existing samples work with
the revised version of the MiP ProMini Pack in this repository. There
will be several follow-up commits to make it a C++ class, more robust,
and more closely match the style of other Arduino libraries.