Refactor out of band notification handling
Previously out of band notifications would be placed in a circular
queue when they were encountered in the serial data stream received
from the MiP. They would only be pulled from this circular queue and
parsed when the user asked for a notification. The code now parses
the out of band notifications as soon as they are encountered in
input serial data stream. The circular queue is no longer required.

NOTE: I did remove the rawReceiveNotification() method which used to
      manually pull notifications from the circular queue since the
      queue no longer exists and they are parsed as soon as they are
      encountered.