Handle MIP_REQUEST_DELAY differently
Previously transportSendRequest() would just call delay() for this
number of milliseconds to make sure that no two requests could be
issued in less time than this. Now it just delays at the top of the
function if this many milliseconds haven't already elapsed since it
was last called. This lets the caller use the CPU to do something
else between calls if they want.