Add a 1k load across 3.3V supply
Without this load, the 3.3V supply was closer to 4.4V. The datasheet
for the regulator indicates that it doesn't have a minimum load to
maintain regulation but it also indicates that the output capacitance
can be closer to 0.5uF when the output current is only 1mA. If a
person was using the 3.3V supply for something else that was putting
a >= 1mA load on it then they could omit this resistor. Lowering the
output capacitance (which was copied from the original Sparkfun design)
might also correct the problem.