Auto switch to 9600 baud for newer MiP robots
The MiP::begin() call will now attempt to first connect at 115200 baud
and then fallback to 9600 baud if that fails. This allows the library
to work with both new and older MiP robots...it just takes a bit longer
to connect with the newer MiP robots while it tries 115200 baud first.