Add unverified LED update APIs
Implements feature request #15. These new APIs allow the chest LED and
head LEDs to be set without the overhead of verifying that the MiP
successfully received and applied the LED state change. Can be useful
if the developer will be setting the LED state multiple times per
second anyway so that a missed state change wouldn't be noticed by a
user.